+++ libcxx/test/std/language.support/support.limits/support.limits.general/bit.version.pass.cpp
@@ -16,6 +16,7 @@
 /*  Constant              Value
     __cpp_lib_bit_cast    201806L [C++2a]
     __cpp_lib_endian      201907L [C++2a]
+    __cpp_lib_int_pow2    202002L [C++2a]
 */
 
 #include <bit>
@@ -31,6 +32,10 @@
 #   error "__cpp_lib_endian should not be defined before c++2a"
 # endif
 
+# ifdef __cpp_lib_int_pow2
+#   error "__cpp_lib_int_pow2 should not be defined before c++2a"
+# endif
+
 #elif TEST_STD_VER == 14
 
 # ifdef __cpp_lib_bit_cast
@@ -41,6 +46,10 @@
 #   error "__cpp_lib_endian should not be defined before c++2a"
 # endif
 
+# ifdef __cpp_lib_int_pow2
+#   error "__cpp_lib_int_pow2 should not be defined before c++2a"
+# endif
+
 #elif TEST_STD_VER == 17
 
 # ifdef __cpp_lib_bit_cast
@@ -49,6 +58,10 @@
 
 # ifdef __cpp_lib_endian
 #   error "__cpp_lib_endian should not be defined before c++2a"
+# endif
+
+# ifdef __cpp_lib_int_pow2
+#   error "__cpp_lib_int_pow2 should not be defined before c++2a"
 # endif
 
 #elif TEST_STD_VER > 17
@@ -71,6 +84,13 @@
 # endif
 # if __cpp_lib_endian != 201907L
 #   error "__cpp_lib_endian should have the value 201907L in c++2a"
+# endif
+
+# ifndef __cpp_lib_int_pow2
+#   error "__cpp_lib_int_pow2 should be defined in c++2a"
+# endif
+# if __cpp_lib_int_pow2 != 202002L
+#   error "__cpp_lib_int_pow2 should have the value 202002L in c++2a"
 # endif
 
 #endif // TEST_STD_VER > 17