+++ libcxx/test/std/numerics/bit/bit.pow.two/bit_floor.pass.cpp
@@ -9,8 +9,9 @@
 // UNSUPPORTED: c++03, c++11, c++14, c++17
 
 // template <class T>
-//   constexpr bool ispow2(T x) noexcept;
+//   constexpr T bit_floor(T x) noexcept;
 
+// Returns: If x == 0, 0; otherwise the maximal value y such that bit_floor(y) is true and y <= x.
 // Remarks: This function shall not participate in overload resolution unless
 //	T is an unsigned integer type
 
@@ -28,15 +29,16 @@ enum class E2 : unsigned char { red };
 template <typename T>
 constexpr bool constexpr_test()
 {
-	return  std::ispow2(T(1))
-	   &&   std::ispow2(T(2))
-	   &&  !std::ispow2(T(3))
-	   &&   std::ispow2(T(4))
-	   &&  !std::ispow2(T(5))
-	   &&  !std::ispow2(T(6))
-	   &&  !std::ispow2(T(7))
-	   &&   std::ispow2(T(8))
-	   &&  !std::ispow2(T(9))
+	return std::bit_floor(T(0)) == T(0)
+	   &&  std::bit_floor(T(1)) == T(1)
+	   &&  std::bit_floor(T(2)) == T(2)
+	   &&  std::bit_floor(T(3)) == T(2)
+	   &&  std::bit_floor(T(4)) == T(4)
+	   &&  std::bit_floor(T(5)) == T(4)
+	   &&  std::bit_floor(T(6)) == T(4)
+	   &&  std::bit_floor(T(7)) == T(4)
+	   &&  std::bit_floor(T(8)) == T(8)
+	   &&  std::bit_floor(T(9)) == T(8)
 	   ;
 }
 
@@ -44,26 +46,26 @@ constexpr bool constexpr_test()
 template <typename T>
 void runtime_test()
 {
-	ASSERT_SAME_TYPE(bool, decltype(std::ispow2(T(0))));
-	ASSERT_NOEXCEPT(                std::ispow2(T(0)));
+	ASSERT_SAME_TYPE(T, decltype(std::bit_floor(T(0))));
+	ASSERT_NOEXCEPT(             std::bit_floor(T(0)));
 
-	assert(!std::ispow2(T(121)));
-	assert(!std::ispow2(T(122)));
-	assert(!std::ispow2(T(123)));
-	assert(!std::ispow2(T(124)));
-	assert(!std::ispow2(T(125)));
-	assert(!std::ispow2(T(126)));
-	assert(!std::ispow2(T(127)));
-	assert( std::ispow2(T(128)));
-	assert(!std::ispow2(T(129)));
-	assert(!std::ispow2(T(130)));
+	assert( std::bit_floor(T(121)) == T(64));
+	assert( std::bit_floor(T(122)) == T(64));
+	assert( std::bit_floor(T(123)) == T(64));
+	assert( std::bit_floor(T(124)) == T(64));
+	assert( std::bit_floor(T(125)) == T(64));
+	assert( std::bit_floor(T(126)) == T(64));
+	assert( std::bit_floor(T(127)) == T(64));
+	assert( std::bit_floor(T(128)) == T(128));
+	assert( std::bit_floor(T(129)) == T(128));
+	assert( std::bit_floor(T(130)) == T(128));
 }
 
 int main(int, char**)
 {
 
     {
-    auto lambda = [](auto x) -> decltype(std::ispow2(x)) {};
+    auto lambda = [](auto x) -> decltype(std::bit_floor(x)) {};
     using L = decltype(lambda);
 
     static_assert( std::is_invocable_v<L, unsigned char>, "");
@@ -145,17 +147,17 @@ int main(int, char**)
 	{
 	__uint128_t val = 128;
 	val <<= 32;
-	assert(!std::ispow2(val-1));
-	assert( std::ispow2(val));
-	assert(!std::ispow2(val+1));
+	assert( std::bit_floor(val-1) == val/2);
+	assert( std::bit_floor(val)   == val);
+	assert( std::bit_floor(val+1) == val);
 	val <<= 2;
-	assert(!std::ispow2(val-1));
-	assert( std::ispow2(val));
-	assert(!std::ispow2(val+1));
+	assert( std::bit_floor(val-1) == val/2);
+	assert( std::bit_floor(val)   == val);
+	assert( std::bit_floor(val+1) == val);
 	val <<= 3;
-	assert(!std::ispow2(val-1));
-	assert( std::ispow2(val));
-	assert(!std::ispow2(val+1));
+	assert( std::bit_floor(val-1) == val/2);
+	assert( std::bit_floor(val)   == val);
+	assert( std::bit_floor(val+1) == val);
 	}
 #endif