+++ libcxx/test/std/numerics/bit/bit.pow.two/bit_width.pass.cpp
@@ -9,15 +9,15 @@
 // UNSUPPORTED: c++03, c++11, c++14, c++17
 
 // template <class T>
-//   constexpr T floor2(T x) noexcept;
+//   constexpr T bit_width(T x) noexcept;
 
-// Returns: If x == 0, 0; otherwise the maximal value y such that floor2(y) is true and y <= x.
+// If x == 0, 0; otherwise one plus the base-2 logarithm of x, with any fractional part discarded.
+
 // Remarks: This function shall not participate in overload resolution unless
 //	T is an unsigned integer type
 
 #include <bit>
 #include <cstdint>
-#include <type_traits>
 #include <cassert>
 
 #include "test_macros.h"
@@ -29,16 +29,16 @@ enum class E2 : unsigned char { red };
 template <typename T>
 constexpr bool constexpr_test()
 {
-	return std::floor2(T(0)) == T(0)
-	   &&  std::floor2(T(1)) == T(1)
-	   &&  std::floor2(T(2)) == T(2)
-	   &&  std::floor2(T(3)) == T(2)
-	   &&  std::floor2(T(4)) == T(4)
-	   &&  std::floor2(T(5)) == T(4)
-	   &&  std::floor2(T(6)) == T(4)
-	   &&  std::floor2(T(7)) == T(4)
-	   &&  std::floor2(T(8)) == T(8)
-	   &&  std::floor2(T(9)) == T(8)
+	return std::bit_width(T(0)) == T(0)
+	   &&  std::bit_width(T(1)) == T(1)
+	   &&  std::bit_width(T(2)) == T(2)
+	   &&  std::bit_width(T(3)) == T(2)
+	   &&  std::bit_width(T(4)) == T(3)
+	   &&  std::bit_width(T(5)) == T(3)
+	   &&  std::bit_width(T(6)) == T(3)
+	   &&  std::bit_width(T(7)) == T(3)
+	   &&  std::bit_width(T(8)) == T(4)
+	   &&  std::bit_width(T(9)) == T(4)
 	   ;
 }
 
@@ -46,26 +46,38 @@ constexpr bool constexpr_test()
 template <typename T>
 void runtime_test()
 {
-	ASSERT_SAME_TYPE(T, decltype(std::floor2(T(0))));
-	ASSERT_NOEXCEPT(             std::floor2(T(0)));
+	ASSERT_SAME_TYPE(T, decltype(std::bit_width(T(0))));
+	ASSERT_NOEXCEPT(             std::bit_width(T(0)));
 
-	assert( std::floor2(T(121)) == T(64));
-	assert( std::floor2(T(122)) == T(64));
-	assert( std::floor2(T(123)) == T(64));
-	assert( std::floor2(T(124)) == T(64));
-	assert( std::floor2(T(125)) == T(64));
-	assert( std::floor2(T(126)) == T(64));
-	assert( std::floor2(T(127)) == T(64));
-	assert( std::floor2(T(128)) == T(128));
-	assert( std::floor2(T(129)) == T(128));
-	assert( std::floor2(T(130)) == T(128));
+	assert( std::bit_width(T(0)) == T(0));
+	assert( std::bit_width(T(1)) == T(1));
+	assert( std::bit_width(T(2)) == T(2));
+	assert( std::bit_width(T(3)) == T(2));
+	assert( std::bit_width(T(4)) == T(3));
+	assert( std::bit_width(T(5)) == T(3));
+	assert( std::bit_width(T(6)) == T(3));
+	assert( std::bit_width(T(7)) == T(3));
+	assert( std::bit_width(T(8)) == T(4));
+	assert( std::bit_width(T(9)) == T(4));
+
+
+	assert( std::bit_width(T(121)) == T(7));
+	assert( std::bit_width(T(122)) == T(7));
+	assert( std::bit_width(T(123)) == T(7));
+	assert( std::bit_width(T(124)) == T(7));
+	assert( std::bit_width(T(125)) == T(7));
+	assert( std::bit_width(T(126)) == T(7));
+	assert( std::bit_width(T(127)) == T(7));
+	assert( std::bit_width(T(128)) == T(8));
+	assert( std::bit_width(T(129)) == T(8));
+	assert( std::bit_width(T(130)) == T(8));
 }
 
 int main(int, char**)
 {
 
     {
-    auto lambda = [](auto x) -> decltype(std::floor2(x)) {};
+    auto lambda = [](auto x) -> decltype(std::bit_width(x)) {};
     using L = decltype(lambda);
 
     static_assert( std::is_invocable_v<L, unsigned char>, "");
@@ -127,6 +139,7 @@ int main(int, char**)
 	static_assert(constexpr_test<__uint128_t>(),        "");
 #endif
 
+
 	runtime_test<unsigned char>();
 	runtime_test<unsigned>();
 	runtime_test<unsigned short>();
@@ -147,17 +160,17 @@ int main(int, char**)
 	{
 	__uint128_t val = 128;
 	val <<= 32;
-	assert( std::floor2(val-1) == val/2);
-	assert( std::floor2(val)   == val);
-	assert( std::floor2(val+1) == val);
+	assert( std::bit_width(val-1) == 39);
+	assert( std::bit_width(val)   == 40);
+	assert( std::bit_width(val+1) == 40);
 	val <<= 2;
-	assert( std::floor2(val-1) == val/2);
-	assert( std::floor2(val)   == val);
-	assert( std::floor2(val+1) == val);
+	assert( std::bit_width(val-1) == 41);
+	assert( std::bit_width(val)   == 42);
+	assert( std::bit_width(val+1) == 42);
 	val <<= 3;
-	assert( std::floor2(val-1) == val/2);
-	assert( std::floor2(val)   == val);
-	assert( std::floor2(val+1) == val);
+	assert( std::bit_width(val-1) == 44);
+	assert( std::bit_width(val)   == 45);
+	assert( std::bit_width(val+1) == 45);
 	}
 #endif